+++ packages/logger-core/src/index.ts
@@ -1,10 +1,11 @@
 export * from './logger.js';
 
-export * from './logger-factory/logger-factory.js';
-export * from './logger-factory/basic-logger-factory.js';
-export * from './logger-factory/stream-logger-factory.js';
+export * from './factory/logger-factory.js';
+export * from './factory/basic-logger-factory.js';
+export * from './factory/stream-logger-factory.js';
 
 export * from './writer/console-writer/console-log-message-writer.js';
 export * from './writer/stream-writer/stream-log-message-writer.js';
 
 export * from './formatter/json-formatter/json-log-message-formatter.js';
+export * from './formatter/json-formatter/json-pretty-log-message-formatter.js';