+++ packages/logger-core/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @corebits/logger-core
 
+## 1.1.0
+
+### Minor Changes
+
+-
+    - Updates LogLevels
+    - Adds LogLevel Silent
+    - Extends Logger with method `close()`
+    - Extends `ILogMessageWriter` interface with method `close()`
+
 ## 1.0.0
 
 ### Major Changes